backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.9
httpx==0.26.0
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.7
pgvector==0.2.4
pandas==2.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
openai==1.12.0
pypdf==4.0.1
email-validator>=2.0.0
